# Fix degrading performance problem