@dasch 2018-06-04
@Elyahou @nguyenquangminh0711 any updates on the performance of the latest version?